rename styles and correct ts errors
