+++ src/components/atoms/Container.ts
@@ -8,6 +8,5 @@ export const Container = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
-
   font-family: "Poppins", sans-serif;
 `;
